# docs/swagger.yaml
definitions:
  api.BMCType:
    enum:
    - dell
    - hpe
    - lenovoxcc
    - lenovoimm
    - nutanix
    type: string
    x-enum-varnames:
    - Dell
    - HPE
    - LenovoXCC
    - LenovoIMM
    - Nutanix
  api.ErrorResponse:
    properties:
      code:
        description: HTTP status code
        example: 400
        type: integer
      message:
        description: Error message
        example: Bad Request
        type: string
    type: object
  api.FirmwareInfoRequest:
    properties:
      bmc_type:
        $ref: '#/definitions/api.BMCType'
      ip_address:
        type: string
    required:
    - ip_address
    type: object
  api.FirmwareUpdateRequest:
    properties:
      bmc_type:
        $ref: '#/definitions/api.BMCType'
      firmware_path:
        description: Path to the firmware file
        type: string
      ip_address:
        type: string
    required:
    - firmware_path
    - ip_address
    type: object
  api.LogType:
    enum:
    - system
    - management
    - fault
    type: string
    x-enum-varnames:
    - System
    - Management
    - Fault
  api.LogsRequest:
    properties:
      bmc_type:
        $ref: '#/definitions/api.BMCType'
      ip_address:
        type: string
      log_type:
        $ref: '#/definitions/api.LogType'
    required:
    - ip_address
    type: object
  api.PowerAction:
    enum:
    - "on"
    - "off"
    - reset
    - bmcreset
    type: string
    x-enum-varnames:
    - "On"
    - "Off"
    - Reset
    - BMCReset
  api.PowerRequest:
    properties:
      action:
        allOf:
        - $ref: '#/definitions/api.PowerAction'
        description: Power action (on, off, reset, bmcreset)
      bmc_type:
        allOf:
        - $ref: '#/definitions/api.BMCType'
        description: Dell, HPE, Lenovo, Nutanix, etc.
      ip_address:
        description: BMC IP address
        type: string
    required:
    - action
    - ip_address
    type: object
  api.PowerResponse:
    properties:
      message:
        example: Power action executed successfully
        type: string
      status:
        description: Status of the action
        example: true
        type: boolean
    type: object
  api.SystemInfoRequest:
    properties:
      bmc_type:
        $ref: '#/definitions/api.BMCType'
      ip_address:
        type: string
    required:
    - ip_address
    type: object
  api.SystemInfoResponse:
    properties:
      biosversion:
        type: string
      cpu: {}
      device: {}
      health:
        type: string
      hostname: {}
      manufacturer:
        type: string
      memory: {}
      model:
        type: string
      powerstate: {}
      responsestatus:
        type: string
      serialnumber:
        type: string
    type: object
info:
  contact: {}
  version: "1.0"
paths:
  /bmc/auth:
    post:
      consumes:
      - application/json
      description: generate JWT token with basic authentication
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Generate Token
      tags:
      - Auth
  /bmc/firmwareinfo:
    post:
      consumes:
      - application/json
      description: This endpoint fetches the firmware info of a BMC device [dell |
        hpe | lenovoxcc | lenovoimm | nutanix].
      parameters:
      - description: Firmware info request parameters
        in: body
        name: firmware_info
        required: true
        schema:
          $ref: '#/definitions/api.FirmwareInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Firmware Info for BMC
      tags:
      - Info
  /bmc/firmwareupdate:
    post:
      consumes:
      - application/json
      deprecated: true
      description: This endpoint updates the firmware of a BMC device.
      parameters:
      - description: Firmware Update request parameters
        in: body
        name: firmware_request
        required: true
        schema:
          $ref: '#/definitions/api.FirmwareUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Firmware Update for BMC
      tags:
      - Patch
  /bmc/logs:
    post:
      consumes:
      - application/json
      description: This endpoint fetches the log info of a BMC device [dell | hpe
        | lenovoxcc | lenovoimm | nutanix].
      parameters:
      - description: log Info request parameters
        in: body
        name: log_info
        required: true
        schema:
          $ref: '#/definitions/api.LogsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Log Info for BMC
      tags:
      - Info
  /bmc/power:
    post:
      consumes:
      - application/json
      description: This endpoint allows the user to run power actions [on | off |
        reset | bmcreset] on a BMC device [dell | hpe | lenovoxcc | lenovoimm | nutanix].
      parameters:
      - description: Power request parameters
        in: body
        name: power_request
        required: true
        schema:
          $ref: '#/definitions/api.PowerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/api.PowerResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Manage power for BMC
      tags:
      - Power
  /bmc/systeminfo:
    post:
      consumes:
      - application/json
      description: This endpoint fetches the system info of a BMC device [dell | hpe
        | lenovoxcc | lenovoimm | nutanix].
      parameters:
      - description: System Info request parameters
        in: body
        name: system_info
        required: true
        schema:
          $ref: '#/definitions/api.SystemInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/api.SystemInfoResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: System Info for BMC
      tags:
      - Info
security:
- BasicAuth: []
  JWT: []
securityDefinitions:
  BasicAuth:
    type: basic
  JWT:
    description: 'JWT Authorization header using the Bearer schema. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"